Simplify Chat to be user friendly, renamed files to main, server & ui
