--- core/gameplay/client/Chat/manifest.json
@@ -1,17 +1,0 @@
-{
-    "schema": 1,
-    "id": "ChatSystem",
-    "networked": true,
-    "version": "0.0.1",
-    "permissions": {
-        "client": {
-            "networking": true,
-            "interface": true,
-            "construct": true,
-            "input": true
-        },
-        "server": {
-            "networking": true
-        }
-    }
-}
